+++ docs/CHANGES.rst
@@ -13,3 +13,5 @@ CHANGELOG
 ------------------
 
 Main version of FeretUI, based on the POC
+
+* Added Response object